✅ QZXN8Q code: split context issue regressions
